refactor : 패키지명 수정 및 파일 정리
