+++ src/main/java/lotto/Model/LottoTicket.java
@@ -1,4 +1,4 @@
-package lottoSecond.Model;
+package lotto.Model;
 
 import java.util.List;
