--- src/main/java/lottoSecond/LottoSecondExecutor.java
@@ -1,10 +1,0 @@
-package lottoSecond;
-
-public class LottoSecondExecutor {
-
-    public static void main(String[] args) {
-        // 로또(2등) 프로그램 시작
-        LottoSecondProgram lottoSecondProgram = new LottoSecondProgram();
-        lottoSecondProgram.run();
-    }
-}
