+++ src/test/java/lotto/Model/LottoTicketTest.java
@@ -1,9 +1,8 @@
-package lottoSecond.Model;
+package lotto.Model;
 
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
-import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
